+++ web/src/App.jsx
@@ -17,7 +17,7 @@ import Share, { loader as shareLoader } from "@/routes/share";
 import ErrorPage from "@/routes/error";
 
 import { SnackbarProvider } from "@/contexts/snackbar/provider";
-import { ThemeProvider } from "@/contexts/theme";
+import { ThemeProvider } from "@/contexts/theme/provider";
 import { UserProvider } from "@/contexts/user";
 import { ConversationProvider } from "@/contexts/conversation/provider";
 import { MessageProvider } from "@/contexts/message/provider";
