+++ web/src/contexts/theme/index.jsx
@@ -1,0 +1,7 @@
+import { createContext } from "react";
+
+
+export const ThemeContext = createContext({
+    theme: "",
+    setTheme: () => { },
+});
